+++ src/onegov.town/onegov/town/models/__init__.py
@@ -5,6 +5,7 @@ from onegov.town.models.file import File, FileCollection
 from onegov.town.models.form import BuiltinFormDefinition, CustomFormDefinition
 from onegov.town.models.image import Image, ImageCollection, Thumbnail
 from onegov.town.models.page import News, Topic
+from onegov.town.models.page_move import PageMove
 from onegov.town.models.resource import DaypassResource
 from onegov.town.models.search import Search
 from onegov.town.models.sitecollection import SiteCollection
@@ -24,6 +25,7 @@ __all__ = [
     'Image',
     'ImageCollection',
     'News',
+    'PageMove',
     'Search',
     'SiteCollection',
     'Thumbnail',